Improved file organisation
